requirements.txt
# FastAPI y servidor
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Base de datos
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Validación y configuración
pydantic==2.10.2
pydantic-settings==2.6.1

# Utilidades
python-dotenv==1.0.1

# Telegram Bot
python-telegram-bot==21.7
